Throw an error if the currently unsupported "rediss" (TLS) protocol is specified
When using a managed Redis database, a `url` with `rediss://` is used for the Redis options, as supported by the latest `ioredis.`

Currently with the legacy `ioredis` dependency, the `url` with `rediss://` option is ignored and the connection defaults silently to `localhost:6379.` The module may appear to work unintentionally if there is a localhost Redis.  

This proposed change will ensure that this module "fails early" and thereby reduce debugging time.